+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [0.0.10] - TBD
 
+### Added
+- Better supports for translation detection and selection
+
 ## [0.0.9] - 2023-08-28
 
 ### Added